The whole application have been rewritten from scratch. It now uses Azauth instead of the old auth library.

+++ application/models/post_model.php
@@ -2,19 +2,29 @@
 
 class Post_model extends Model
 {
-	function __construct()
+	// constructor
+	public function __construct()
 	{
 		parent::Model();
 	}
 	
+	// php4 compatibility
+	public function Post_model()
+	{
+		self::__construct();
+	}
+	
 	/**
-	 * Retrieve many posts
+	 * Retrieves all the posts from the database.
+	 * The first parameter is used to be able to
+	 * see inactive posts in the admin posts area.
 	 *
-	 * @param int $all_posts
+	 * @access public
+	 * @param int $all_posts 
 	 * @param int $num 
 	 * @return array
-	 */
-	function get_posts($all_posts = NULL, $num = NULL) // NOTE: The first parameter is used to be able to see inactive posts in the admin posts area
+	 **/
+	public function get_posts($all_posts = NULL, $num = NULL)
 	{
 		$this->db->select('posts.id, posts.title, posts.body, posts.created_at, posts.updated_at, posts.active, users.username');
 		$this->db->from('posts');
@@ -33,16 +43,19 @@ class Post_model extends Model
 		
 		// if the query did not retrieve any data, return NULL
 		return NULL;
-	}
+	} // End of get_posts
 	
 	/**
-	 * Get one post
+	 * Get one post from the database
+	 * The second parameter is used to be able
+	 * to edit inactive posts when logged in
 	 *
+	 * @access public
 	 * @param int $post_id 
-	 * @param int $is_active
+	 * @param int $is_active 
 	 * @return object
-	 */
-	function get_post($post_id, $is_active = 1) // NOTE: the second parameter is used to be able to edit inactive posts when logged in
+	 **/
+	public function get_post($post_id, $is_active = 1)
 	{
 		$this->db->select('posts.id, posts.title, posts.body, posts.created_at, posts.updated_at, posts.active, users.username');
 		$this->db->from('posts');
@@ -59,31 +72,38 @@ class Post_model extends Model
 		
 		// if the query did not retrieve a row, return NULL
 		return NULL;
-	}
+	} // End of get_post
+	
+// =================
+// = Admin Methods =
+// =================
 	
 	/**
-	 * Inserts a new post in the database
+	 * Inserts new post in the database.
 	 *
-	 * @param array $post
-	 * @return void
+	 * @access public
+	 * @param string $post 
+	 * @return bool
 	 **/
-	function new_post($post)
+	public function new_post($post)
 	{
 		// if the post variable provided is empty, return FALSE...
 		if(empty($post))
 			return FALSE;
 		// ...otherwise, insert a new post in the database
 		$this->db->insert('posts',$post);
-	}
+		return TRUE;
+	} // End of new_post
 	
 	/**
 	 * Update an existing post
 	 *
+	 * @access public
 	 * @param int $post_id 
 	 * @param array $post 
 	 * @return void
 	 **/
-	function update_post($post_id,$post)
+	public function update_post($post_id,$post)
 	{
 		// if the post variable provided is empty, return FALSE...
 		if(empty($post))
@@ -91,10 +111,31 @@ class Post_model extends Model
 		// ...otherwise, update the correct post in the database
 		$this->db->where('id',$post_id);
 		$this->db->update('posts',$post);
-	}
+	} // End of update_post
 	
-	function delete_post($post_id)
+	/**
+	 * Deletes one post from the database.
+	 *
+	 * @todo send the user to a confirmation page before deleting a post
+	 *
+	 * @access public
+	 * @param int $post_id 
+	 * @return bool
+	 **/
+	public function delete_post($post_id)
 	{
+		if (!is_valid_number($post_id))
+			return FALSE;
+		
 		$this->db->delete('posts',array('id' => $post_id));
-	}
-}
+		return TRUE;
+	} // End of delete_post
+	
+// ------------------------
+// - End of Admin Methods -
+// ========================
+	
+} // End of Post_model
+
+/* End of file post_model.php */
+/* Location: ./application/model/post_model.php */
